Include exception to deletecategory with salon id

+++ category-service/src/main/java/com/meerthika/service/impl/CategoryServiceImpl.java
@@ -44,8 +44,12 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategoryById(Long id) throws Exception {
+    public void deleteCategoryById(Long id, Long salonId) throws Exception {
         Category category = getCategoryById(id);
+
+        if(category.getSalonId().equals(salonId)){
+            throw new Exception("You dont have permission to delete this category");
+        }
         categoryRepository.deleteById(id);
     }
 }
